Fixed the provider= method.
Set the default resource group for all methods to use the armrest_configuration.resource_group method.

Commented out the create and update methods, since they aren't actually implemented yet.

Explicitly raise an error in most methods if a resource group is not provided.

Refactored the list method slightly, only creating a mutex if there's a result.

Flip-flop the arguments to the capture method to be in line with other service classes, and allow the resource_group to be specified as an option.

Allow user defined resource group to override default in capture method.