Use @hashobj for resource_group & subscription_id
The "id" field is almost an expected field to exist for most results
coming back from azure, but both the `BaseModel#resource_group` and
`BaseModel#subscription_id` methods require that a `.id` method exists,
which will be defined in `__setobj__`.

To allow those methods to work with `skip_accessors_definition` set to
true when initializing and `BaseModel` class, the `#id_from_hash` method
was introduced to support that.  It checks for either a `"id"` or a
`:id` to exists in the `@hashobj`.  This new method is then used in
place of the `id` dynamic method.