Use ArmrestCollection#next_link in get_all_results
Currently, when the #get_all_results method is called, it parses the
response body (as JSON) once when instantiating the ArmrestCollection,
and a second time to determine the `'nextLink'` value for the subsequent
`while` loop to get the remaining results (this is duplicated as well
for each iteration of the loop).

Since we now have a `next_link` variable on the ArmrestCollection
instances that is determined in the same fashion from the first
`JSON.parse`, we can use that instead of duplicating the same effort
twice.

In practice, this halved the allocations caused by the JSON gem (when
using the one from ruby stdlib), and 10% of the total object allocations
over all when calling `StorageAccountService#list_all_private_images` on
a small to moderate collection of `StorageAccounts` (around 200 in this
instance, filtered by location to around 70).  CPU improvements might be
observed on larger collections.