Add BaseModel::attr_from_hash private class method
Used as a standardized way of defining a common pattern of methods to be
used for light model instantiation.

This allows exposing top level key methods on class declaration (instead
of object instantiation) which can be used by the azure-armrest
internals as a common way working with intermediate models, and avoiding
expensive add_accessor_methods for each object instantiation.