Relax nokogiri gem version
There is currently a CVE out for `libxml2`, that affects `nokogiri`:

https://access.redhat.com/security/cve/cve-2017-9050

Unfortunately, so projects that depend on `azure-armrest` expect that
the version of nokogiri is within the `1.7.x` release.  To allow for
that (as insecure as it is), but make sure we try to use the latest
version of `nokogiri` when possible, relax the version requirement on
this gem so that we are secure as we can be without breaking other
projects.

That said, we probably should make nokogiri-1.8.1 a minimum requirement
in the v1.0 release though.