Use caller_locations instead of caller
Ruby already has a mechanism in place for getting the callstack path
locations and line numbers that doesn't require string
splitting/parsing.

This also fixes a bug on Windows systems when trying to split callstacks
with file paths in them, since most windows file paths include something
like 'C:/path/to/file', which has an additional ':' that is not expected
by the codebase.