Fix StorageAccountService#list_private_images
When speeding up the `list_all_private_images` method, the `list_all`
method was updated with the skip_accessors_definition flag to that and
associated methods.  This was not, though, applied to the
`#list_private_images` method, which works on just a single storage
account at a time.

This updates `#list` in both the StorageAccountService version, and the
base class to allow it to support this flag.