Managed Disk Read Semantics Change
Azure Managed Disk Instances have been taking exceptionally long in
some locations.  It turns out that the semantic that azure-armrest
uses to perform reads via the get_blob_raw method is the culprit.
The method gets a SAS URL - an access token to the disk - each
time it performs a read operation.  The call to get the access token
is asynchronous and if it does not return immediately with success when
poll is called the gem sleeps for 5 seconds.  This adds 5 seconds to
*every* call to get data.

get_blob_raw is being left in the gem for compatibility purposes.
However the new semantic is this:

Call open on the service (disk service or snapshot service).
Open returns a ManagedDisk object.
Call read on the ManagedDisk object as many times as you need, then call
close on the ManagedDisk object when finished.

Under the covers the ManagedDisk will contain the SAS URL to be used.  It should
only be obtained once unless it times out, in which case the caller is expected
to get a new ManagedDisk object to continue.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1508154